more wrapping, first test accessing RAW and image data working

+++ setup.py
@@ -83,22 +83,24 @@ def windows_libraw_compile():
     # download cmake to compile libraw
     # the cmake zip contains a cmake-3.0.1-win32-x86 folder when extracted
     cmake_url = 'http://www.cmake.org/files/v3.0/cmake-3.0.1-win32-x86.zip'
-    cmake = os.path.abspath('external/cmake-3.0.1-win32-x86/bin/cmake')
-    files = [(cmake_url, 'cmake-3.0.1-win32-x86.zip', 'external')]
-    for url, path, extractdir in files:
-        if not os.path.exists(path):
-            print('Downloading', url)
-            urlretrieve(url, path)
-        with zipfile.ZipFile(path) as z:
-            print('Extracting', path, 'into', extractdir)
-            z.extractall(extractdir)
+    cmake = os.path.abspath('external/cmake-3.0.1-win32-x86/bin/cmake.exe')
+    files = [(cmake_url, 'cmake-3.0.1-win32-x86.zip', 'external', cmake)]
+    for url, path, extractdir, extractcheck in files:
+        if not os.path.exists(extractcheck):
+            if not os.path.exists(path):
+                print('Downloading', url)
+                urlretrieve(url, path)
+        
+            with zipfile.ZipFile(path) as z:
+                print('Extracting', path, 'into', extractdir)
+                z.extractall(extractdir)
     
     # configure and compile libraw
     cwd = os.getcwd()
     if not os.path.exists(cmake_build):
         os.mkdir(cmake_build)
     os.chdir(cmake_build)
-    cmds = [cmake + ' .. -G "NMake Makefiles" -DENABLE_EXAMPLES=OFF -DENABLE_RAWSPEED=OFF',
+    cmds = [cmake + ' .. -G "NMake Makefiles" -DENABLE_EXAMPLES=OFF -DENABLE_OPENMP=OFF -DENABLE_RAWSPEED=OFF',
             'dir',
             'nmake raw_r' # build only thread-safe version ('raw'=non-thread-safe)
             ]
